Minor (moved a few classes)
