+++ src/Ably.Core/Ably.Core.csproj
@@ -54,7 +54,7 @@
     <Compile Include="CapabilityResource.cs" />
     <Compile Include="CipherParams.cs" />
     <Compile Include="Config.cs" />
-    <Compile Include="ConnectionDetailsMessage.cs" />
+    <Compile Include="Types\ConnectionDetailsMessage.cs" />
     <Compile Include="CustomSerialisers\CapabilityJsonConverter.cs" />
     <Compile Include="CustomSerialisers\CapabilityMessagePackSerializer.cs" />
     <Compile Include="CustomSerialisers\DateTimeOffsetJsonConverter.cs" />
@@ -74,11 +74,11 @@
     <Compile Include="IAblyHttpClient.cs" />
     <Compile Include="IAuthCommands.cs" />
     <Compile Include="IChannelCipher.cs" />
-    <Compile Include="IEncodedMessage.cs" />
+    <Compile Include="Types\IEncodedMessage.cs" />
     <Compile Include="IPaginatedResource.cs" />
     <Compile Include="IRestClient.cs" />
     <Compile Include="Logger.cs" />
-    <Compile Include="Message.cs" />
+    <Compile Include="Types\Message.cs" />
     <Compile Include="MessageEncoders\Base64Encoder.cs" />
     <Compile Include="MessageEncoders\CipherEncoder.cs" />
     <Compile Include="MessageEncoders\IMessageHandler.cs" />
@@ -91,7 +91,7 @@
     <Compile Include="Platform\ICrypto.cs" />
     <Compile Include="Platform\IoC.cs" />
     <Compile Include="Platform\IPlatform.cs" />
-    <Compile Include="PresenceMessage.cs" />
+    <Compile Include="Types\PresenceMessage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Protocol.cs" />
     <Compile Include="QueryDirection.cs" />
