+++ src/Ably.Core/Types/IEncodedMessage.cs
